Certificate Generation Service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
passlib[bcrypt]
python-multipart
